refactor(1013): remove WiredClientManufacturerReportGenerator facade (SC-001, position 26)
Extract WiredClientManufacturerReportGenerator (~128 LOC) from MistHelper.py monolith
to src/reports/wired_client_manufacturer_report_generator.py following the SC-001
facade pattern. Menu handler that fetches wired clients org-wide, exports the full
set, then optionally exports a filtered subset by user-selected manufacturer.

- New: src/reports/wired_client_manufacturer_report_generator.py (9 static methods,
  lazy `mh = importlib.import_module("MistHelper")` for ConfigUtils/apisession/
  InputUtils/DataProcessingUtils/DataExporter access)
- MistHelper.py: add alphabetical re-export alias in src.reports.* import block;
  replace class body with breadcrumb comment (not stub -- preserves alias)
- pyproject.toml: add coverage omit entry (alphabetical placement)

Local gate green: black + ruff + py_compile + `python MistHelper.py --test`
baseline 66/0/131 preserved.
